+++ mirrord-layer/src/file/ops.rs
@@ -2,8 +2,8 @@ use std::{ffi::CString, io::SeekFrom, os::unix::io::RawFd, path::PathBuf};
 
 use libc::{c_int, c_uint, FILE, O_CREAT, O_RDONLY, S_IRUSR, S_IWUSR, S_IXUSR};
 use mirrord_protocol::{
-    CloseFileResponse, OpenFileResponse, OpenOptionsInternal, ReadFileResponse, ResponseError,
-    SeekFileResponse, WriteFileResponse,
+    CloseFileResponse, OpenFileResponse, OpenOptionsInternal, ReadFileResponse, SeekFileResponse,
+    WriteFileResponse,
 };
 use tokio::sync::oneshot;
 use tracing::{debug, error};
@@ -80,8 +80,7 @@ fn close_remote_file_on_failure(fd: usize) -> Result<CloseFileResponse, LayerErr
     // Close the remote file if the call to `libc::shm_open` failed and we have an invalid local fd.
     error!("Call to `libc::shm_open` resulted in an error, closing the file remotely!");
 
-    let (file_channel_tx, file_channel_rx) =
-        oneshot::channel::<Result<CloseFileResponse, ResponseError>>();
+    let (file_channel_tx, file_channel_rx) = oneshot::channel();
 
     blocking_send_file_message(HookMessageFile::Close(Close {
         fd,
@@ -100,8 +99,7 @@ pub(crate) fn openat(
         "openat -> trying to open valid file {:?} with relative dir {:?}.",
         path, relative_fd
     );
-    let (file_channel_tx, file_channel_rx) =
-        oneshot::channel::<Result<OpenFileResponse, ResponseError>>();
+    let (file_channel_tx, file_channel_rx) = oneshot::channel();
 
     let open_options = OpenOptionsInternalExt::from_flags(open_flags);
 
@@ -188,8 +186,7 @@ pub(crate) fn fdopen(
 pub(crate) fn read(fd: usize, read_amount: usize) -> Result<ReadFileResponse, LayerError> {
     debug!("read -> trying to read valid file {:?}.", fd);
 
-    let (file_channel_tx, file_channel_rx) =
-        oneshot::channel::<Result<ReadFileResponse, ResponseError>>();
+    let (file_channel_tx, file_channel_rx) = oneshot::channel();
 
     let reading_file = Read {
         fd,
@@ -205,8 +202,7 @@ pub(crate) fn read(fd: usize, read_amount: usize) -> Result<ReadFileResponse, La
 
 pub(crate) fn lseek(fd: usize, seek_from: SeekFrom) -> Result<u64, LayerError> {
     debug!("lseek -> trying to seek valid file {:?}.", fd);
-    let (file_channel_tx, file_channel_rx) =
-        oneshot::channel::<Result<SeekFileResponse, ResponseError>>();
+    let (file_channel_tx, file_channel_rx) = oneshot::channel();
 
     let seeking_file = Seek {
         fd,
@@ -222,8 +218,7 @@ pub(crate) fn lseek(fd: usize, seek_from: SeekFrom) -> Result<u64, LayerError> {
 
 pub(crate) fn write(fd: usize, write_bytes: Vec<u8>) -> Result<isize, LayerError> {
     debug!("write -> trying to write valid file {:?}.", fd);
-    let (file_channel_tx, file_channel_rx) =
-        oneshot::channel::<Result<WriteFileResponse, ResponseError>>();
+    let (file_channel_tx, file_channel_rx) = oneshot::channel();
 
     let writing_file = Write {
         fd,
@@ -239,8 +234,7 @@ pub(crate) fn write(fd: usize, write_bytes: Vec<u8>) -> Result<isize, LayerError
 
 pub(crate) fn close(fd: usize) -> Result<c_int, LayerError> {
     debug!("close -> trying to close valid file {:?}.", fd);
-    let (file_channel_tx, file_channel_rx) =
-        oneshot::channel::<Result<CloseFileResponse, ResponseError>>();
+    let (file_channel_tx, file_channel_rx) = oneshot::channel();
 
     let closing_file = Close {
         fd,