+++ mirrord-protocol/src/codec.rs
@@ -189,13 +189,16 @@ pub struct WriteFileResponse {
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub struct CloseFileResponse;
 
+/// Type alias for `Result`s that should be returned from mirrord-agent to mirrord-layer.
+pub type RemoteResult<T> = Result<T, ResponseError>;
+
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub enum FileResponse {
-    Open(Result<OpenFileResponse, ResponseError>),
-    Read(Result<ReadFileResponse, ResponseError>),
-    Seek(Result<SeekFileResponse, ResponseError>),
-    Write(Result<WriteFileResponse, ResponseError>),
-    Close(Result<CloseFileResponse, ResponseError>),
+    Open(RemoteResult<OpenFileResponse>),
+    Read(RemoteResult<ReadFileResponse>),
+    Seek(RemoteResult<SeekFileResponse>),
+    Write(RemoteResult<WriteFileResponse>),
+    Close(RemoteResult<CloseFileResponse>),
 }
 
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
@@ -252,9 +255,6 @@ impl From<dns_lookup::AddrInfo> for AddrInfoInternal {
     }
 }
 
-#[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
-pub struct GetAddrInfoResponse(pub Result<Vec<AddrInfoInternal>, ResponseError>);
-
 /// `-agent` --> `-layer` messages.
 #[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
 pub enum DaemonMessage {
@@ -263,8 +263,8 @@ pub enum DaemonMessage {
     LogMessage(LogMessage),
     File(FileResponse),
     Pong,
-    GetEnvVarsResponse(Result<HashMap<String, String>, ResponseError>),
-    GetAddrInfoResponse(GetAddrInfoResponse),
+    GetEnvVarsResponse(RemoteResult<HashMap<String, String>>),
+    GetAddrInfoResponse(RemoteResult<Vec<AddrInfoInternal>>),
 }
 
 pub struct ClientCodec {